[Move] 기존 패키지 service, utils 좀 더 세분화 및 이동
