+++ NoMoneyTrip/app/src/main/java/com/mtjin/nomoneytrip/service/notification/NotificationBody.kt
@@ -1,3 +1,5 @@
 package com.mtjin.nomoneytrip.service
 
+import com.mtjin.nomoneytrip.service.notification.NotificationData
+
 data class NotificationBody(val to: String, val data: NotificationData)
