+++ NoMoneyTrip/app/src/main/java/com/mtjin/nomoneytrip/service/work_manager/ScheduledWorker.kt
@@ -1,4 +1,4 @@
-package com.mtjin.nomoneytrip.service
+package com.mtjin.nomoneytrip.service.work_manager
 
 import android.content.Context
 import android.util.Log
